// packages/expo-styling/src/types.ts
export type ExtractedStyleValue =
  | string
  | number
  | RuntimeValue
  | ExtractedStyleValue[];
// ...
export type InteropMeta = {
  /* The processed styled props */
  styledProps: Record<string, Style>;
  /* The processed styled props */
  styledPropsMeta: Record<string, PropInteropMeta>;
  /* Inline variables */
  variables: Record<string, unknown>;
  /* Inline container runtime info */
  containers: Record<string, ContainerRuntime>;
  /* A list of props that will be processed by useAnimated  */
  animatedProps: Set<string>;
  /* A list of props that will be processed by useTransition  */
  transitionProps: Set<string>;
  /* Does this component need the onLayout handler  */
  requiresLayout: boolean;
  /* Does variables have any values  */
  hasInlineVariables: boolean;
  /* Does variables have any values  */
  hasInlineContainers: boolean;
  /* If defined, we should wrap in the AnimationInterop */
  animationInteropKey?: string;
};

export type PropInteropMeta = {
  variables?: Record<string, unknown>;
  containers?: string[];
  animated: boolean;
  transition: boolean;
  requiresLayout: boolean;
};
// ...
export interface MutableSignal<T = unknown> {
  get(): T;
  set(value: T): void;
}

export type Interaction = {
  active: MutableSignal<boolean>;
  hover: MutableSignal<boolean>;
  focus: MutableSignal<boolean>;
  layout: {
    width: MutableSignal<number>;
    height: MutableSignal<number>;
  };
};

export type ExtractedContainer = {
  names?: string[];
  type: ContainerType;
};

export type ContainerRuntime = {
  type: ContainerType;
  interaction: Interaction;
  style: Style;
};
// ...
export type ExtractedAnimation = {
  frames: ExtractedKeyframe[];
  requiresLayout?: boolean;
};

export type ExtractedKeyframe = {
  selector: number;
  style: Record<string, ExtractedStyleValue>;
};

export type PseudoClassesQuery = {
  hover?: boolean;
  active?: boolean;
  focus?: boolean;
};

export type StyleSheetRegisterOptions = {
  declarations?: Record<string, ExtractedStyle | ExtractedStyle[]>;
  keyframes?: Record<string, ExtractedAnimation>;
};

export type Style = ViewStyle | TextStyle | ImageStyle;
export type StyleProp = Style | StyleProp[] | undefined;